drv:src: Sync API changes between main_aig and release/main_aig branch (#486)

+++ driver/src/dma/xaie_dma.h
@@ -85,14 +85,14 @@ XAIE_AIG_EXPORT AieRC XAie_DmaSetAddrOffsetLen(XAie_DmaDesc *DmaDesc, XAie_MemIn
 			       u64 Offset, u32 Len);
 XAIE_AIG_EXPORT AieRC XAie_DmaSetMultiDimAddr(XAie_DmaDesc *DmaDesc, XAie_DmaTensor *Tensor,
 		u64 Addr, u32 Len);
-XAIE_AIG_EXPORT AieRC XAie_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u8 Wrap,
+XAIE_AIG_EXPORT AieRC XAie_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u16 Wrap,
 		u8 IterCurr);
 XAIE_AIG_EXPORT AieRC XAie_DmaSetPadding(XAie_DmaDesc *DmaDesc, XAie_DmaPadTensor *PadTensor);
 XAIE_AIG_EXPORT AieRC XAie_DmaEnableCompression(XAie_DmaDesc *DmaDesc);
 XAIE_AIG_EXPORT AieRC XAie_DmaConfigFifoMode(XAie_DmaDesc *DmaDesc,
 		XAie_DmaFifoCounter Counter);
 XAIE_AIG_EXPORT AieRC XAie_DmaGetNumBds(XAie_DevInst *DevInst, XAie_LocType Loc, u8 *NumBds);
-XAIE_AIG_EXPORT AieRC XAie_DmaSetNextBd(XAie_DmaDesc *DmaDesc, u8 NextBd, u8 EnableNextBd);
+XAIE_AIG_EXPORT AieRC XAie_DmaSetNextBd(XAie_DmaDesc *DmaDesc, u16 NextBd, u8 EnableNextBd);
 XAIE_AIG_EXPORT AieRC XAie_DmaEnableBd(XAie_DmaDesc *DmaDesc);
 XAIE_AIG_EXPORT AieRC XAie_DmaDisableBd(XAie_DmaDesc *DmaDesc);
 XAIE_AIG_EXPORT AieRC XAie_DmaSetAxi(XAie_DmaDesc *DmaDesc, u8 Smid, u8 BurstLen, u8 Qos,
@@ -100,9 +100,9 @@ XAIE_AIG_EXPORT AieRC XAie_DmaSetAxi(XAie_DmaDesc *DmaDesc, u8 Smid, u8 BurstLen
 XAIE_AIG_EXPORT AieRC XAie_DmaSetInterleaveEnable(XAie_DmaDesc *DmaDesc, u8 DoubleBuff,
 		u8 IntrleaveCount, u16 IntrleaveCurr);
 XAIE_AIG_EXPORT AieRC XAie_DmaWriteBd(XAie_DevInst *DevInst, XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 XAIE_AIG_EXPORT AieRC XAie_DmaReadBd(XAie_DevInst *DevInst, XAie_DmaDesc *DmaDesc,
-		XAie_LocType Loc, u8 BdNum);
+		XAie_LocType Loc, u16 BdNum);
 XAIE_AIG_EXPORT AieRC XAie_DmaChannelResetAll(XAie_DevInst *DevInst, XAie_LocType Loc,
 		XAie_DmaChReset Reset);
 XAIE_AIG_EXPORT AieRC XAie_DmaChannelReset(XAie_DevInst *DevInst, XAie_LocType Loc,
@@ -112,7 +112,7 @@ XAIE_AIG_EXPORT AieRC XAie_DmaChannelPauseStream(XAie_DevInst *DevInst, XAie_Loc
 XAIE_AIG_EXPORT AieRC XAie_DmaChannelPauseMem(XAie_DevInst *DevInst, XAie_LocType Loc, u8 ChNum,
 		XAie_DmaDirection Dir, u8 Pause);
 XAIE_AIG_EXPORT AieRC XAie_DmaChannelPushBdToQueue(XAie_DevInst *DevInst, XAie_LocType Loc,
-		u8 ChNum, XAie_DmaDirection Dir, u8 BdNum);
+		u8 ChNum, XAie_DmaDirection Dir, u16 BdNum);
 XAIE_AIG_EXPORT AieRC XAie_DmaChannelEnable(XAie_DevInst *DevInst, XAie_LocType Loc, u8 ChNum,
 		XAie_DmaDirection Dir);
 XAIE_AIG_EXPORT AieRC XAie_DmaChannelDisable(XAie_DevInst *DevInst, XAie_LocType Loc, u8 ChNum,
@@ -130,7 +130,7 @@ XAIE_AIG_EXPORT AieRC XAie_DmaGetPendingBdCount(XAie_DevInst *DevInst, XAie_LocT
 XAIE_AIG_EXPORT AieRC XAie_DmaGetMaxQueueSize(XAie_DevInst *DevInst, XAie_LocType Loc,
 		u8 *QueueSize);
 XAIE_AIG_EXPORT AieRC XAie_DmaChannelSetStartQueue(XAie_DevInst *DevInst, XAie_LocType Loc,
-		u8 ChNum, XAie_DmaDirection Dir, u8 BdNum, u32 RepeatCount,
+		u8 ChNum, XAie_DmaDirection Dir, u16 BdNum, u32 RepeatCount,
 		u8 EnTokenIssue);
 XAIE_AIG_EXPORT AieRC XAie_DmaChannelSetStartQueueGeneric(XAie_DevInst *DevInst,
 		XAie_LocType Loc, u8 ChNum, XAie_DmaDirection Dir,
@@ -153,11 +153,11 @@ XAIE_AIG_EXPORT AieRC XAie_DmaSetZeroPadding(XAie_DmaDesc *DmaDesc, u8 Dim,
 XAIE_AIG_EXPORT AieRC XAie_DmaTlastEnable(XAie_DmaDesc *DmaDesc);
 XAIE_AIG_EXPORT AieRC XAie_DmaTlastDisable(XAie_DmaDesc *DmaDesc);
 XAIE_AIG_EXPORT AieRC XAie_DmaGetBdLen(XAie_DevInst *DevInst, XAie_LocType Loc, u32 *Len,
-		u8 BdNum);
+		u16 BdNum);
 XAIE_AIG_EXPORT AieRC XAie_DmaUpdateBdLen(XAie_DevInst *DevInst, XAie_LocType Loc, u32 Len,
-		u8 BdNum);
+		u16 BdNum);
 XAIE_AIG_EXPORT AieRC XAie_DmaUpdateBdAddr(XAie_DevInst *DevInst, XAie_LocType Loc, u64 Addr,
-		u8 BdNum);
+		u16 BdNum);
 XAIE_AIG_EXPORT AieRC XAie_DmaSetPadValue(XAie_DevInst *DevInst, XAie_LocType Loc, u8 ChNum,
 		u32 PadValue);
 XAIE_AIG_EXPORT AieRC XAie_DmaGetChannelStatus(XAie_DevInst *DevInst, XAie_LocType Loc,
